Extracted the active_record adapter models to another file, and lazy load them
