+++ lib/flipper-active_record.rb
@@ -1,5 +1,6 @@
 require 'active_support/lazy_load_hooks'
+require 'flipper/adapters/active_record'
 
 ActiveSupport.on_load(:active_record) do
-  require 'flipper/adapters/active_record'
+  require 'flipper/adapters/active_record/models'
 end
