+++ lib/flipper/adapters/active_record.rb
@@ -6,24 +6,6 @@ module Flipper
     class ActiveRecord
       include ::Flipper::Adapter
 
-      # Private: Do not use outside of this adapter.
-      class Feature < ::ActiveRecord::Base
-        self.table_name = [
-          ::ActiveRecord::Base.table_name_prefix,
-          "flipper_features",
-          ::ActiveRecord::Base.table_name_suffix,
-        ].join
-      end
-
-      # Private: Do not use outside of this adapter.
-      class Gate < ::ActiveRecord::Base
-        self.table_name = [
-          ::ActiveRecord::Base.table_name_prefix,
-          "flipper_gates",
-          ::ActiveRecord::Base.table_name_suffix,
-        ].join
-      end
-
       # Public: The name of the adapter.
       attr_reader :name
